% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ref_Draft_Scraper.R
\name{Ref_Draft_Scraper}
\alias{Ref_Draft_Scraper}
\title{HockeyRef Draft URL Scraper}
\usage{
Ref_Draft_Scraper(website, ages = c(17, 50), playerStats = "all",
  goalieStats = "all", Season = "R", sepTeam = F)
}
\arguments{
\item{website}{the url of the draft year webpage on hockey-reference.com.}

\item{ages}{2-length vector for the age ranges desired. first number is lower-bound, second number is upper-bound. This will be applied to all players.}

\item{playerStats}{vector of the wanted player stats. possible values:
stateSep - seperate goals, assists, and points based on strength (PP, SH, EV)
Cor = Corsi,
Fen = Fenwick,
oiS = On-Ice shooting and save percentage,
PDO = PDO,
IceTime = Ice time,
Awards = Awards with placement,
pmBreak = plus/minus breakdown (GF, GA, etc.),
PS = HockeyRef's point shares metric,
xGF = Expected Goals For percentage}

\item{goalieStats}{vector of the wanted goalie stats. possible values:
QS = quality starts metric
GSAA = goals saved above average
Scoring = goals/assists/points/etc.
Awards - Awards with placement}

\item{Season}{Determines if regular season data, playoff data, or both will be returned for all the players. Currently only 'R' works.}

\item{sepTeam}{Boolean about whether you want rows for every team a player played on in the NHL in a give year, or if you just want the cumulative production combining the two teams.}
}
\value{
data frame (or list of data frames) with data from all the players drafted in the given year.
}
\description{
Returns data from all players in the given draft year url
}
