% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ref_Goalie_Scraper.R
\name{Ref_Goalie_Scraper}
\alias{Ref_Goalie_Scraper}
\title{HockeyRef Player Scraper}
\usage{
Ref_Goalie_Scraper(website, ages = c(17, 50), Stats, Season = "R",
  sepTeam = F)
}
\arguments{
\item{website}{the url of the goalie's webpage on hockey-reference.com.}

\item{ages}{2-length vector for the age ranges desired. first number is lower-bound, second number is upper-bound.}

\item{Season}{Determines if regular season data, playoff data, or both will be returned for all the players. 'R' for Regular Season, 'P' for Playoffs, 'RP' for both Regular Season and Playoffs}

\item{sepTeam}{Boolean about whether you want rows for every team a goalie played on in the NHL in a give year, or if you just want the cumulative production combining the two teams.}

\item{goalieStats}{vector of the wanted goalie stats. possible values:
QS = quality starts metric
GSAA = goals saved above average
Scoring = goals/assists/points/etc.
Awards - Awards with placement}
}
\value{
data frame (or list of data frames) with data the desired goalie.
}
\description{
Returns data from a player given their url.
}
