% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ref_Player_Scraper.R
\name{Ref_Player_Scraper}
\alias{Ref_Player_Scraper}
\title{HockeyRef Player Scraper}
\usage{
Ref_Player_Scraper(website, ages = c(17, 50), Stats = "all",
  Season = "R", sepTeam = F)
}
\arguments{
\item{website}{the url of the player's webpage on hockey-reference.com.}

\item{ages}{2-length vector for the age ranges desired. first number is lower-bound, second number is upper-bound.}

\item{Season}{Determines if regular season data, playoff data, or both will be returned for all the players. 'R' for Regular Season, 'P' for Playoffs, 'RP' for both Regular Season and Playoffs}

\item{sepTeam}{Boolean about whether you want rows for every team a player played on in the NHL in a give year, or if you just want the cumulative production combining the two teams.}

\item{playerStats}{vector of the wanted player stats. possible values:
stateSep - seperate goals, assists, and points based on strength (PP, SH, EV)
Cor = Corsi,
Fen = Fenwick,
oiS = On-Ice shooting and save percentage,
PDO = PDO,
IceTime = Ice time,
Awards = Awards with placement,
pmBreak = plus/minus breakdown (GF, GA, etc.),
PS = HockeyRef's point shares metric,
xGF = Expected Goals For percentage}
}
\value{
data frame (or list of data frames) with data the desired player.
}
\description{
Returns data from a player given their url.
}
