% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_rbind.R
\name{smart_rbind}
\alias{smart_rbind}
\title{Smarter rbind}
\usage{
smart_rbind(table1, table2)
}
\arguments{
\item{table1}{data.frame}

\item{table2}{data.frame}
}
\value{
stacked combo of both tables. Is able to deal with tables of different columns
}
\description{
Returns a stacked table that combines two tables. Can combine tables with different number of columns.
}
